requirements.txt
pymupdf4llm
llama_index
azure-search-documents
openai
langchain
langgraph
fastapi
uvicorn